progress clean up PF node implementation
