--- mrpt_pf_localization/include/mrpt_localization/mrpt_localization_defaults.h
@@ -1,14 +1,0 @@
-/* +------------------------------------------------------------------------+
-   |                             mrpt_navigation                            |
-   |                                                                        |
-   | Copyright (c) 2014-2023, Individual contributors, see commit authors   |
-   | See: https://github.com/mrpt-ros-pkg/mrpt_navigation                   |
-   | All rights reserved. Released under BSD 3-Clause license. See LICENSE  |
-   +------------------------------------------------------------------------+ */
-
-#pragma once
-#define MRPT_LOCALIZATION_DEFAULT_DEBUG true
-#define MRPT_LOCALIZATION_DEFAULT_GUI_MRPT true
-#define MRPT_LOCALIZATION_DEFAULT_INI_FILE "pf-localization.ini"
-#define MRPT_LOCALIZATION_DEFAULT_MAP_FILE ""
-#define MRPT_LOCALIZATION_DEFAULT_SENSOR_SOURCES "scan,scan1,scan2"
