--- mrpt_pf_localization/include/mrpt_localization_node_defaults.h
@@ -1,17 +1,0 @@
-/* +------------------------------------------------------------------------+
-   |                             mrpt_navigation                            |
-   |                                                                        |
-   | Copyright (c) 2014-2023, Individual contributors, see commit authors   |
-   | See: https://github.com/mrpt-ros-pkg/mrpt_navigation                   |
-   | All rights reserved. Released under BSD 3-Clause license. See LICENSE  |
-   +------------------------------------------------------------------------+ */
-
-#ifndef MRPT_LOCALIZATION_NODE_DEFAULTS_H
-#define MRPT_LOCALIZATION_NODE_DEFAULTS_H
-
-#define MRPT_LOCALIZATION_NODE_DEFAULT_RATE 10.0
-#define MRPT_LOCALIZATION_NODE_DEFAULT_PARAMETER_UPDATE_SKIP 1
-#define MRPT_LOCALIZATION_NODE_DEFAULT_PARTICLECLOUD_UPDATE_SKIP 5
-#define MRPT_LOCALIZATION_NODE_DEFAULT_MAP_UPDATE_SKIP 2
-
-#endif	// MRPT_LOCALIZATION_NODE_DEFAULTS_H
